deployement/requirement.txt
numpy
tensorflow
pillow
matplotlib
streamlit
scipy
